// frontend/src/lib/api/model/matchParticipantResponseDto.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * Student Arena API
 * API Documentation
 * OpenAPI spec version: 1.0
 */
import type { MatchParticipantResponseDtoRole } from './matchParticipantResponseDtoRole';

export interface MatchParticipantResponseDto {
  /** Match participant ID */
  id: number;
  /** Player ID */
  playerId: number;
  /** Player game name with tagline */
  playerName: string;
  /** Player PUUID */
  puuid: string;
  /** Champion ID */
  championId: number;
  /** Player role in the match */
  role: MatchParticipantResponseDtoRole;
  /** Team ID (100 or 200) */
  teamId: number;
  /**
   * Kills
   * @nullable
   */
  kills?: number | null;
  /**
   * Deaths
   * @nullable
   */
  deaths?: number | null;
  /**
   * Assists
   * @nullable
   */
  assists?: number | null;
  /**
   * Creep score
   * @nullable
   */
  cs?: number | null;
  /**
   * Gold earned
   * @nullable
   */
  goldEarned?: number | null;
  /**
   * Item IDs
   * @nullable
   */
  items?: number[] | null;
  /**
   * Summoner spell IDs
   * @nullable
   */
  spells?: number[] | null;
}